Update mainwindow_interface.py
Now DC widget allows adjustment of "Pshift" parameter for "time" based data (time,observation,weight). This feature can be used in the case of light curve modelling of eclipsing binaries on eccentric orbit. However, T0 and Pshift is perfectly correlated if the period is constant, and extremely highly correlated if the period is not constant (see the source code of the DC program at ftp://ftp.astro.ufl.edu/pub/wilson/lcdc2015/). Extreme caution is advised if both parameters will be adjusted simultaneously.